+++ src/main/java/com/example/green/domain/challenge/repository/TeamChallengeRepositoryImpl.java
@@ -139,7 +139,7 @@ public class TeamChallengeRepositoryImpl implements TeamChallengeRepositoryCusto
 			challenge.getBeginDateTime(),
 			challenge.getEndDateTime(),
 			challenge.getChallengeImage(),
-			challenge.getChallengePoint().getAmount().intValue()
+			challenge.getChallengePoint().getAmount()
 		);
 	}
 }
